Move js library files to lib/
