+++ src/js/lib/message.js
