+++ src/js/lib/snapshot.js
